server/requirements.txt
Flask==2.2.2
Flask-API==3.0.post1
python-logging-loki==0.3.1
pymongo==4.3.3
